chore(.gitignore): add .env to .gitignore to prevent accidental commit of sensitive data feat(__root.tsx): lazy load TanStackRouterDevtools to improve initial load performance
The .env file is added to .gitignore to prevent sensitive data like API keys, database credentials, etc., from being accidentally committed to the repository.

In __root.tsx, TanStackRouterDevtools is now lazy-loaded, which means it will only be loaded when it's needed. This change improves the initial load performance of the application, especially in development mode. Also, it will not be loaded in production mode, further optimizing the performance.